A Logo now cannot be deleted if it is in use (this utilizes a new method I put in the model to find out if in use)

+++ app/Http/Controllers/LogoController.php
@@ -221,7 +221,15 @@ class LogoController extends Controller
         
         if(isset($logo))
         {
-            $this->deleteLogo($logo, $this->logoDirectory);
+            if(!$logo->inUse())
+            {
+                $this->deleteLogo($logo, $this->logoDirectory);
+            }
+            else
+            {
+                $errorText = "Unable to delete logo, as it is currently in use";
+                return redirect()->back()->with('customErrors', [$errorText])->withInput();
+            }
         }
         else
         {
